Convert DlpContentRestriction and SystemFeature enum to enum class.
Otherwise, if both are included than there is kMaxValue collision.
Also, enum classes are preferable over just enums.

Bug: 1311441
Test: Trybots.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3560459
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987036}

+++ chrome/browser/apps/app_service/publishers/extension_apps_chromeos.cc
@@ -696,7 +696,7 @@ void ExtensionAppsChromeOs::OnSystemFeaturesPrefChanged() {
   is_disabled_apps_mode_hidden_ = is_pref_disabled_mode_hidden;
 
   UpdateAppDisabledState(disabled_system_features_pref,
-                         policy::SystemFeature::kWebStore,
+                         static_cast<int>(policy::SystemFeature::kWebStore),
                          extensions::kWebStoreAppId, is_disabled_mode_changed);
 }
