+++ chrome/browser/ash/web_applications/scanning_app_integration_browsertest.cc
@@ -36,7 +36,7 @@ IN_PROC_BROWSER_TEST_P(ScanningAppIntegrationTest, ScanningAppDisabled) {
     ListPrefUpdate update(TestingBrowserProcess::GetGlobal()->local_state(),
                           policy::policy_prefs::kSystemFeaturesDisableList);
     base::Value* list = update.Get();
-    list->Append(policy::SystemFeature::kScanning);
+    list->Append(static_cast<int>(policy::SystemFeature::kScanning));
   }
 
   ASSERT_FALSE(GetManager()
