+++ chrome/browser/chrome_content_browser_client_unittest.cc
@@ -519,7 +519,7 @@ TEST_F(ChromeContentSettingsRedirectTest, RedirectSettingsURL) {
   EXPECT_EQ(settings_url, dest_url);
 
   base::Value list(base::Value::Type::LIST);
-  list.Append(policy::SystemFeature::kBrowserSettings);
+  list.Append(static_cast<int>(policy::SystemFeature::kBrowserSettings));
   testing_local_state_.Get()->Set(
       policy::policy_prefs::kSystemFeaturesDisableList, std::move(list));
 
@@ -537,7 +537,7 @@ TEST_F(ChromeContentSettingsRedirectTest, RedirectOSSettingsURL) {
   EXPECT_EQ(os_settings_url, dest_url);
 
   base::Value list(base::Value::Type::LIST);
-  list.Append(policy::SystemFeature::kOsSettings);
+  list.Append(static_cast<int>(policy::SystemFeature::kOsSettings));
   testing_local_state_.Get()->Set(
       policy::policy_prefs::kSystemFeaturesDisableList, std::move(list));
 
@@ -560,7 +560,7 @@ TEST_F(ChromeContentSettingsRedirectTest, RedirectScanningAppURL) {
   EXPECT_EQ(scanning_app_url, dest_url);
 
   base::Value list(base::Value::Type::LIST);
-  list.Append(policy::SystemFeature::kScanning);
+  list.Append(static_cast<int>(policy::SystemFeature::kScanning));
   testing_local_state_.Get()->Set(
       policy::policy_prefs::kSystemFeaturesDisableList, std::move(list));
 
@@ -577,7 +577,7 @@ TEST_F(ChromeContentSettingsRedirectTest, RedirectCameraAppURL) {
   EXPECT_EQ(camera_app_url, dest_url);
 
   base::Value list(base::Value::Type::LIST);
-  list.Append(policy::SystemFeature::kCamera);
+  list.Append(static_cast<int>(policy::SystemFeature::kCamera));
   testing_local_state_.Get()->Set(
       policy::policy_prefs::kSystemFeaturesDisableList, std::move(list));
 
@@ -594,7 +594,7 @@ TEST_F(ChromeContentSettingsRedirectTest, RedirectHelpURL) {
   EXPECT_EQ(GURL("chrome://settings/help"), dest_url);
 
   base::Value list(base::Value::Type::LIST);
-  list.Append(policy::SystemFeature::kBrowserSettings);
+  list.Append(static_cast<int>(policy::SystemFeature::kBrowserSettings));
   testing_local_state_.Get()->Set(
       policy::policy_prefs::kSystemFeaturesDisableList, std::move(list));
