+++ chrome/browser/web_applications/extensions/web_app_policy_manager_unittest.cc
@@ -1112,7 +1112,7 @@ TEST_P(WebAppPolicyManagerTest, DisableWebApps) {
   // Add camera to system features disable list policy.
   auto disabled_apps_list =
       std::make_unique<base::Value>(base::Value::Type::LIST);
-  disabled_apps_list->Append(policy::SystemFeature::kCamera);
+  disabled_apps_list->Append(static_cast<int>(policy::SystemFeature::kCamera));
   testing_local_state_.Get()->SetUserPref(
       policy::policy_prefs::kSystemFeaturesDisableList,
       std::move(disabled_apps_list));
